requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
schedule>=1.2.1
firecrawl-py>=1.0.0
typing-extensions>=4.9.0
pytz>=2024.1
pydantic>=2.0.0
volcengine-python-sdk[ark]>=1.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.2
newsapi-python>=0.2.6
